crypto_agent/requirements.txt
python-binance==1.0.19
pandas==2.2.3
numpy==2.1.3
python-dotenv==1.0.1
openai==1.54.3
requests==2.32.3
httpx==0.27.2
langchain-ollama>=0.2.0
langgraph>=0.2.0
